test(test): merge form post callback parameters
