+++ packages/integration-tests/src/client/index.ts
@@ -6,6 +6,7 @@ import ky, { type KyInstance } from 'ky';
 
 import { demoAppRedirectUri, logtoUrl } from '#src/constants.js';
 
+import { getSubmittingCallbackUri } from './callback-uri.js';
 import { MemoryStorage } from './storage.js';
 
 export const defaultConfig = {
@@ -55,37 +56,6 @@ const isCookiePathMatched = (cookiePath: string, requestPath: string) =>
   cookiePath === '/' ||
   requestPath === cookiePath ||
   requestPath.startsWith(cookiePath.endsWith('/') ? cookiePath : `${cookiePath}/`);
-
-const decodeHtmlAttribute = (value: string) =>
-  value
-    .replaceAll('&quot;', '"')
-    .replaceAll('&#34;', '"')
-    .replaceAll('&apos;', "'")
-    .replaceAll('&#39;', "'")
-    .replaceAll('&amp;', '&');
-
-const getHtmlAttribute = (html: string, attribute: string) =>
-  new RegExp(`\\b${attribute}=(["'])(.*?)\\1`, 'i').exec(html)?.[2];
-
-const getSubmittingCallbackUri = (html: string, redirectUri?: string) => {
-  const form = /<form\b[^>]*>/i.exec(html)?.[0];
-  assert(form, new Error('Missing callback form'));
-  const action = getHtmlAttribute(form, 'action');
-  assert(action, new Error('Missing callback form action'));
-  const callbackUrl = new URL(redirectUri ?? decodeHtmlAttribute(action));
-  const callbackParameters = new URLSearchParams(callbackUrl.search);
-  for (const [input] of html.matchAll(/<input\b[^>]*>/gi)) {
-    const name = getHtmlAttribute(input, 'name');
-    const value = getHtmlAttribute(input, 'value');
-
-    if (name && value) {
-      callbackParameters.set(decodeHtmlAttribute(name), decodeHtmlAttribute(value));
-    }
-  }
-  const search = callbackParameters.toString();
-
-  return `${callbackUrl.origin}${callbackUrl.pathname}${search && `?${search}`}${callbackUrl.hash}`;
-};
 
 export default class MockClient {
   public rawCookies: string[] = [];
@@ -205,7 +175,6 @@ export default class MockClient {
     const authResponseLocation = authResponse.headers.get('location');
 
     if (authResponse.status === 200) {
-      const body = await authResponse.text();
       const signInSession: unknown = JSON.parse(
         (await this.storage.getItem(PersistKey.SignInSession)) ?? 'null'
       );
@@ -216,7 +185,7 @@ export default class MockClient {
         typeof signInSession.redirectUri === 'string'
           ? signInSession.redirectUri
           : undefined;
-      const signInCallbackUri = getSubmittingCallbackUri(body, redirectUri);
+      const signInCallbackUri = getSubmittingCallbackUri(await authResponse.text(), redirectUri);
 
       this.mergeRawCookies(authResponse.headers.getSetCookie());
       await this.logto.handleSignInCallback(signInCallbackUri);
